separate main from package
